+++ infrastructure-provisioning/src/general/scripts/os/rstudio_install_dataengine_kernels.py
@@ -19,15 +19,11 @@
 # ******************************************************************************
 
 import argparse
-import os
-
-import backoff
 from fabric.api import *
 from fabric.contrib.files import exists
-
 from dlab.meta_lib import *
+import os
 from dlab.fab import *
-
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--cluster_name', type=str, default='')
@@ -40,33 +36,28 @@ parser.add_argument('--notebook_ip', type=str, default='')
 parser.add_argument('--datalake_enabled', type=str, default='false')
 args = parser.parse_args()
 
-@backoff.on_exception(backoff.expo, SystemExit, max_time=200)
-def try_put():
-    try:
-        put(scripts_dir + 'rstudio_dataengine_create_configs.py',
-            '/tmp/rstudio_dataengine_create_configs.py')
-        put(templates_dir + 'notebook_spark-defaults_local.conf',
-            '/tmp/notebook_spark-defaults_local.conf')
-    except:
-        sys.exit(1)
-
 
 def configure_notebook(keyfile, hoststring):
     scripts_dir = '/root/scripts/'
     templates_dir = '/root/templates/'
-    try_put()
+    run('mkdir -p /tmp/{}/'.format(args.cluster_name))
+    if not exists('/tmp/rstudio_dataengine_create_configs.py'):
+        put(scripts_dir + 'rstudio_dataengine_create_configs.py', '/tmp/rstudio_dataengine_create_configs.py')
+    put(templates_dir + 'notebook_spark-defaults_local.conf', '/tmp/{}/notebook_spark-defaults_local.conf'.format(args.cluster_name))
     spark_master_ip = args.spark_master.split('//')[1].split(':')[0]
     spark_memory = get_spark_memory(True, args.os_user, spark_master_ip, keyfile)
-    run('echo "spark.executor.memory {0}m" >> /tmp/notebook_spark-defaults_local.conf'.format(spark_memory))
-    sudo('\cp /tmp/rstudio_dataengine_create_configs.py /usr/local/bin/rstudio_dataengine_create_configs.py')
-    sudo('chmod 755 /usr/local/bin/rstudio_dataengine_create_configs.py')
-    sudo('mkdir -p /usr/lib/python2.7/dlab/')
-    run('mkdir -p /tmp/dlab_libs/')
-    local('scp -i {} /usr/lib/python2.7/dlab/* {}:/tmp/dlab_libs/'.format(keyfile, hoststring))
-    run('chmod a+x /tmp/dlab_libs/*')
-    sudo('mv /tmp/dlab_libs/* /usr/lib/python2.7/dlab/')
-    if exists('/usr/lib64'):
-        sudo('ln -fs /usr/lib/python2.7/dlab /usr/lib64/python2.7/dlab')
+    run('echo "spark.executor.memory {0}m" >> /tmp/{1}/notebook_spark-defaults_local.conf'.format(spark_memory, args.cluster_name))
+    if not exists('/usr/local/bin/rstudio_dataengine_create_configs.py'):
+        sudo('\cp /tmp/rstudio_dataengine_create_configs.py /usr/local/bin/rstudio_dataengine_create_configs.py')
+        sudo('chmod 755 /usr/local/bin/rstudio_dataengine_create_configs.py')
+    if not exists('/usr/lib/python2.7/dlab/'):
+        sudo('mkdir -p /usr/lib/python2.7/dlab/')
+        run('mkdir -p /tmp/dlab_libs/')
+        local('scp -i {} /usr/lib/python2.7/dlab/* {}:/tmp/dlab_libs/'.format(keyfile, hoststring))
+        run('chmod a+x /tmp/dlab_libs/*')
+        sudo('cp -r /tmp/dlab_libs/* /usr/lib/python2.7/dlab/')
+        if exists('/usr/lib64'):
+            sudo('ln -fs /usr/lib/python2.7/dlab /usr/lib64/python2.7/dlab')
 
 
 if __name__ == "__main__":
